이슈 조회 api 폴더 이동 (issueDetail -> global)
