+++ webFrontend/src/pages/issueDetail/api/hooks.ts
@@ -1,23 +1,5 @@
-import { useMutation, useQuery } from "@tanstack/react-query";
-import { deletePostLike, getIssueDetail, postLike } from "./func";
-
-/**
- * 상세 이슈 조회 훅
- */
-export const useGetIssue = ({
-  id,
-  enabled,
-}: {
-  id: string;
-  enabled: boolean;
-}) => {
-  const { data, isLoading, refetch } = useQuery({
-    queryKey: ["issue", id],
-    queryFn: () => getIssueDetail({ id }),
-    enabled: enabled,
-  });
-  return { issueData: data, isLoadingIssue: isLoading, refetchIssue: refetch };
-};
+import { useMutation } from "@tanstack/react-query";
+import { deletePostLike, postLike } from "./func";
 
 /**
  * 상세 이슈 도움돼요 추가 훅
